Reintegrating the changes chris needs to finish the encounter mechanic.
