+++ SoteriaGame/Assets/Scenes/TileEventSystem.unity
@@ -513,6 +513,7 @@ GameObject:
   - 23: {fileID: 619391266}
   - 54: {fileID: 619391267}
   - 65: {fileID: 619391265}
+  - 114: {fileID: 619391272}
   - 114: {fileID: 619391271}
   - 114: {fileID: 619391270}
   m_Layer: 0
@@ -596,36 +597,45 @@ Transform:
 --- !u!114 &619391270
 MonoBehaviour:
   m_ObjectHideFlags: 0
-  m_PrefabParentObject: {fileID: 11431584, guid: f9505a09d1fc0e44cb3a06d9d6781bcb,
-    type: 2}
-  m_PrefabInternal: {fileID: 630415047}
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
   m_GameObject: {fileID: 619391264}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 0b3b9f26208477e4aa52436115d162fc, type: 3}
+  m_Script: {fileID: 11500000, guid: f1a785edea3a3804cad611161f24e0f5, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  numSoteriaStatues: 1
+  soteriaStatuePrefab: {fileID: 181762, guid: fef1a5f1b3dd1c04a937e708a1370ce1, type: 2}
+  fearResidue: 0
+  playerRotation: 0
+  forcedRotation: 0
+  gameOverTimer: 30
+  Coins: 0
 --- !u!114 &619391271
 MonoBehaviour:
   m_ObjectHideFlags: 0
-  m_PrefabParentObject: {fileID: 11431586, guid: f9505a09d1fc0e44cb3a06d9d6781bcb,
-    type: 2}
-  m_PrefabInternal: {fileID: 630415047}
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 619391264}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 09dd6d3416d679246a4881fc8ee050c1, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  moveSpeed: 1
+  playerRotation: 0
+--- !u!114 &619391272
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
   m_GameObject: {fileID: 619391264}
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: f6a5fdffe9b6bc84a968daf3a9cda2ce, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  moveSpeed: .5
-  numSoteriaStatues: 5
-  soteriaStatuePrefab: {fileID: 181762, guid: fef1a5f1b3dd1c04a937e708a1370ce1, type: 2}
-  fearResidue: 0
-  playerRotation: -10
-  forcedRotation: .25
-  gameOverTimer: 10
-  Coins: 0
-  currentState: 0
 --- !u!1001 &630415047
 Prefab:
   m_ObjectHideFlags: 0
@@ -681,7 +691,9 @@ Prefab:
       propertyPath: m_LocalPosition.x
       value: 39.7999992
       objectReference: {fileID: 0}
-    m_RemovedComponents: []
+    m_RemovedComponents:
+    - {fileID: 11431586, guid: f9505a09d1fc0e44cb3a06d9d6781bcb, type: 2}
+    - {fileID: 11431584, guid: f9505a09d1fc0e44cb3a06d9d6781bcb, type: 2}
   m_ParentPrefab: {fileID: 100100000, guid: f9505a09d1fc0e44cb3a06d9d6781bcb, type: 2}
   m_RootGameObject: {fileID: 619391264}
   m_IsPrefabParent: 0
