+++ SoteriaGame/Assets/Scripts/Controllers/EncounterMovementController.cs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f1a785edea3a3804cad611161f24e0f5
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
